requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
